this next commit contains an updated project 4
dan the lab man asked me to update my project 4 and re-submit it via
git. i think the problem was that it was not working in a browser even
though the code was mostly correct. i just had to spend a few hours
debugging it but now its ready fo re-submission. here is the
re-submission.